requirements.txt
pytest
python-dateutil
pylint
pytest_cov
requests
requests_mock
json